我的基金，有复盘数据时涨幅计算错误;开盘前，百分比显示正负有问题

+++ src/com/dkhs/portfolio/ui/StockQuotesActivity.java
@@ -596,8 +596,8 @@ public class StockQuotesActivity extends ModelAcitivity implements OnClickListen
             }
             tvPercentage.setText(StringFromatUtils.get2PointPercentPlus(mStockQuotesBean.getPercentage()));
             tvHuanShouLv.setText(StringFromatUtils.get2PointPercent(mStockQuotesBean.getTurnover_rate() * 100));
-            tvChengjiaoLiang.setText(StringFromatUtils.convertToWanHand((int) mStockQuotesBean.getVolume()));
-            tvChengjiaoE.setText(StringFromatUtils.convertToWan((int) mStockQuotesBean.getAmount()));
+            tvChengjiaoLiang.setText(StringFromatUtils.convertToWanHand(mStockQuotesBean.getVolume()));
+            tvChengjiaoE.setText(StringFromatUtils.convertToWan(mStockQuotesBean.getAmount()));
         }
 
         if (isIndexType()) {
